update cli flow
