+++ src/cli/mod.rs
@@ -1,3 +1,3 @@
-pub mod tui;
+pub mod cli;
 pub mod open;
 pub mod exit;
