--- src/tui/open.rs
@@ -1,8 +1,0 @@
-pub fn open() -> String {
-    // TODO:
-    // octocrab?? use github api, get list of users repos, append public|private to each
-    // put in list
-    // logic to open repo, editor, tmux, etc....
-    "open function!!!!".to_string()
-}
-
